feat: created the basic layout of the form app
In the layout, the progress bar, the brand logo and the typeForm
redirecting FAB is created.

CSS has been reset and a few CSS variables are created for use. While
few files have been deleted that are currently unused, a complete
cleanup is necessary in a separate commit.

A button utility class has also been created. Though, it is expected to
be removed once we make a separe button component.